backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.0
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.0
python-dotenv==1.0.0
httpx==0.25.2
websockets==12.0
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1